+++ grpc_client_impl.go
@@ -6,6 +6,7 @@ import (
 	aggregatev1 "github.com/chalk-ai/chalk-go/gen/chalk/aggregate/v1"
 	commonv1 "github.com/chalk-ai/chalk-go/gen/chalk/common/v1"
 	"github.com/chalk-ai/chalk-go/gen/chalk/engine/v1/enginev1connect"
+	serverv1 "github.com/chalk-ai/chalk-go/gen/chalk/server/v1"
 	"github.com/chalk-ai/chalk-go/gen/chalk/server/v1/serverv1connect"
 	"github.com/chalk-ai/chalk-go/internal"
 	"github.com/cockroachdb/errors"
@@ -38,7 +39,9 @@ func newGrpcClient(cfg ClientConfig) (*grpcClientImpl, error) {
 	if logger == nil {
 		logger = DefaultLeveledLogger
 	}
-	config.getToken = MakeGetTokenFunc(logger, authClient)
+	config.getToken = func(clientId string, clientSecret string) (*getTokenResult, error) {
+		return getToken(clientId, clientSecret, logger, authClient)
+	}
 
 	// Necessary to get GRPC engines URL
 	if err := config.refresh(false); err != nil {
@@ -60,7 +63,35 @@ func newGrpcClient(cfg ClientConfig) (*grpcClientImpl, error) {
 	}, nil
 }
 
-func (c *grpcClientImpl) UpdateAggregates(args UpdateAggregatesParams) (*commonv1.UploadFeaturesBulkResponse, error) {
+func getToken(clientId string, clientSecret string, logger LeveledLogger, client serverv1connect.AuthServiceClient) (*getTokenResult, error) {
+	logger.Debugf("Getting new token via gRPC")
+	authRequest := connect.NewRequest(
+		&serverv1.GetTokenRequest{
+			ClientId:     clientId,
+			ClientSecret: clientSecret,
+			GrantType:    "client_credentials",
+		},
+	)
+	token, err := client.GetToken(context.Background(), authRequest)
+	if err != nil {
+		logger.Debugf("Failed to get a new token: %s", err.Error())
+		return nil, err
+	}
+
+	expiresAt := token.Msg.GetExpiresAt()
+	if token.Msg.GetExpiresAt() == nil {
+		return nil, errors.New("token has no expiration date")
+	}
+
+	return &getTokenResult{
+		ValidUntil:         expiresAt.AsTime(),
+		AccessToken:        token.Msg.GetAccessToken(),
+		PrimaryEnvironment: token.Msg.GetPrimaryEnvironment(),
+		Engines:            token.Msg.GetGrpcEngines(),
+	}, nil
+}
+
+func (c *grpcClientImpl) UpdateAggregates(ctx context.Context, args UpdateAggregatesParams) (*commonv1.UploadFeaturesBulkResponse, error) {
 	inputsConverted, err := getConvertedInputsMap(args.Inputs)
 	if err != nil {
 		return nil, wrapClientError(err, "error converting inputs map")
@@ -74,11 +105,6 @@ func (c *grpcClientImpl) UpdateAggregates(args UpdateAggregatesParams) (*commonv
 		InputsFeather: inputsFeather,
 		BodyType:      commonv1.FeatherBodyType_FEATHER_BODY_TYPE_TABLE,
 	})
-
-	ctx := context.Background()
-	if args.Context != nil {
-		ctx = args.Context
-	}
 
 	res, err := c.queryClient.UploadFeaturesBulk(ctx, req)
 	if err != nil {
